+++ src/seed/tests/loader/cssjs/a.js
@@ -1,0 +1,3 @@
+KISSY.add("a",function(){
+
+});